Build #283 -- Version 0.0.0.18322 ---------------------------------
Directory trunk:

COMMON.C:
DEBUG.C:
DISPDBG.C:
EXCEPT.C:
FREE.C:
IMMEXEC.C:
MAIN.H:
PF_DOMINO.C:
RESDEBUG.C:
SC_ERASE.C:
SC_SAVE.C:
STARTADDR.H:
  Prepare for 64-bit compiles.

CONVERT.C:
  Use <OptionFlags> to hold global Boolean data.

DATATYPE.H:
  Move <PtrReusedXXX> macros to <macros.h>.
  Prepare for 64-bit compiles.

DFNHDR.Y:
  Omit debugging messages from default display.
  Use TRUE/FALSE instead of 0/1 whenever Boolean.

EDIT.RC:
  Define Paste menu item for PC/3270.

EDITCTRL.H:
  Prepare for 64-bit compiles.
  Use only one <MYWM_PASTE_APL> message #.

EDITFCN.C:
  Inlcude UNISCRIBE definition (if any).
  Split apart the pieces of <Unitrans.h> into <unitranshdr.h>
    (header defintions only) and <unitranstab.h> (table defintions only).
  Remove from ToDo list items already done.
  Prepare for 64-bit compiles.
  Use only one <MYWM_PASTE_APL> message #.
  Zap <lpLclMemVirtStr> ptr in WM_DESTROY in case we re-enter this message.
  Handle UTF16_A_ to UTF16_Z_ with Alternate Font.
  Use <OptionFlags> to hold global Boolean data.
  Optionally, translate UTF16_A_ through UTF_Z_ to lowercase.

EXECDFN.C:
GETFNS.C:
PF_DTACKJOT.C:
PF_QUERY.C:
PRIMFNS.C:
QF_EX.C:
QF_NC.C:
QF_SIZE.C:
QF_SYSVER.C:
STRAND.C:
SYMTAB.C:
SYSVARS.C:
TOKENIZE.C:
TYPEMOTE.C:
VER.C:
YYFNS.C:
  Prepare for 64-bit compiles.
  Use TRUE/FALSE instead of 0/1 whenever Boolean.

EXTERNS.H:
  Define default global options.
  Move previous separate global options into one struc.
  Mark UNISCRIBE-sensitive code as such.

MACROS.H:
  Move <PtrReusedXXX> macros from <datatype.h> to here.

MAIN.C:
  Include UNISCRIBE definition (if any) for <CreateNewFontSM>.
  Include Unicode Translation Header for <MYWM_PASTE_APL> arg.
  Mark UNISCRIBE-sensitive code as such.
  Prepare for 64-bit compiles.
  Use only one <MYWM_PASTE_APL> message #.

MAKEFILE.W32:
MAKEFILE.W64:
  Split <MAKEFILE> into two parts:  32- and 64-bit.

MK32.BAT:
MKD32.BAT:
MK64.BAT:
MKD64.BAT:
  Split <MK.BAT> and <mkd.bat> into two parts:  32- and 64-bit

PARSE.Y:
  Omit debugging messages from default display.
  Prepare for 64-bit compiles.

PERTAB.H:
  Move <UNISCRIBE> define to <uniscribe.h>.

QF_TS.C:
  Use TRUE/FALSE instead of 0/1 whenever Boolean.
  Use <OptionFlags> to hold global Boolean data.

RESOURCE.H:
  Define Paste menu item for PC/3270.

SAVEFCN.C:
  Prepare for 64-bit compiles.
  Reorganize <SF_xxx> functions.
  Fix Bug:  Define new function <SF_NumLinesM> for []FX char matrix args
    instead of <SF_NumLinesCom> as we need to return the # rows not cols.

SAVEFCN.H:
  Remove #define which associates <SF_NumLinesM> with common function
    as we need a separate function for that now.

SC_LOAD.C:
  Fix bug:  Surround code in <LoadWorkspace_EM> with __try/__except
    so we catch Page Faults in uncommitted memory.
  Use <OptionFlags> to hold global Boolean data.

SESSMAN.C:
  Include UNISCRIBE definition (if any) for call to <CoInitialize>/<CoUnintialize>.
  Prepare for 64-bit compiles.
  Use only one <MYWM_PASTE_APL> message #.

SIS.H:
  Improve comments.

SYSCMDS.C:
  Use <OptionFlags> to hold global Boolean data.

TABCTRL.C:
  Prepare for 64-bit compiles.
  Mark DRAWTABNUM-specific code as such.

UNICODE.H:
  Define codes for A underbar through Z underbar
    using SImPL's Private Use Area.

UNISCRIBE.H:
  Use separate file for definition of <UNISCRIBE>.

UNITRANSHDR.H:
UNITRANSTAB.H:
  Split apart the pieces of <Unitrans.h> into <unitranshdr.h>
    (header defintions only) and <unitranstab.h> (table defintions only).


Directory trunk\WineHQ:

EDITCTRL.C:
  Prepare for 64-bit compiles.
  Recall parent with EN_MAXTEXT notification as necessary.
  Define action for PC/2370 paste.


Directory trunk\WineHQ\Include:

BASETSD.H:
WIN.H:
WINBASE.H:
  Prepare for 64-bit compiles.


Directory trunk\WineHQ\Include\msvcrt\Wine:

DEBUG.H:
  Prepare for 64-bit compiles.

+++ pf_dtackjot.c
@@ -380,7 +380,7 @@ LPPL_YYSTYPE PrimFnMonDownTackJot_EM_YY
                 // Fill in the result token
                 lpYYRes->tkToken.tkFlags.TknType   = TKT_VARIMMED;
                 lpYYRes->tkToken.tkFlags.ImmType   = IMMTYPE_CHAR;
-////////////////lpYYRes->tkToken.tkFlags.NoDisplay = 0;     // Already zero from YYAlloc
+////////////////lpYYRes->tkToken.tkFlags.NoDisplay = FALSE; // Already zero from YYAlloc
                 lpYYRes->tkToken.tkData.tkChar     = aplCharRht;
                 lpYYRes->tkToken.tkCharIndex       = lptkFunc->tkCharIndex;
 
@@ -567,7 +567,7 @@ NORMAL_EXIT:
     // Fill in the result token
     lpYYRes->tkToken.tkFlags.TknType   = TKT_VARARRAY;
 ////lpYYRes->tkToken.tkFlags.ImmType   = 0;     // Already zero from YYAlloc
-////lpYYRes->tkToken.tkFlags.NoDisplay = 0;     // Already zero from YYAlloc
+////lpYYRes->tkToken.tkFlags.NoDisplay = FALSE; // Already zero from YYAlloc
     lpYYRes->tkToken.tkData.tkGlbData  = MakePtrTypeGlb (hGlbRes);
     lpYYRes->tkToken.tkCharIndex       = lptkFunc->tkCharIndex;
 
@@ -793,7 +793,7 @@ LPAPLCHAR CompileArrInteger
             lpFmtColStr[aplDimCol].uLdBl = max (lpFmtColStr[aplDimCol].uLdBl, uLen);
 
             // Calculate the length of the integer part
-            uLen = (lpaplChar - lpwszOut) - 1;
+            uLen = (UINT) (lpaplChar - lpwszOut) - 1;
 
             // Max the current integer width with this
             lpFmtColStr[aplDimCol].uInts = max (lpFmtColStr[aplDimCol].uInts, uLen);
@@ -915,14 +915,14 @@ LPAPLCHAR CompileArrFloat
             if (lpwsz)
             {
                 // Calculate the length of the integer part
-                uLen = (lpwsz - lpwszOut);
+                uLen = (UINT) (lpwsz - lpwszOut);
 
                 // Max the current integer width with this
                 lpFmtColStr[aplDimCol].uInts = max (lpFmtColStr[aplDimCol].uInts, uLen);
 
                 // Calculate the length of the fractional part
                 //   including the decimal point
-                uLen = (lpaplChar - lpwsz) - 1;
+                uLen = (UINT) (lpaplChar - lpwsz) - 1;
 
                 // Max the current fractional width with this
                 lpFmtColStr[aplDimCol].uFrcs = max (lpFmtColStr[aplDimCol].uFrcs, uLen);
@@ -932,7 +932,7 @@ LPAPLCHAR CompileArrFloat
             } else  // No decimal point
             {
                 // Calculate the length of the integer part
-                uLen = (lpaplChar - lpwszOut) - 1;
+                uLen = (UINT) (lpaplChar - lpwszOut) - 1;
 
                 // Max the current integer width with this
                 lpFmtColStr[aplDimCol].uInts = max (lpFmtColStr[aplDimCol].uInts, uLen);
@@ -1233,7 +1233,7 @@ LPAPLCHAR CompileArrAPA
             lpFmtColStr[aplDimCol].uLdBl = max (lpFmtColStr[aplDimCol].uLdBl, uLen);
 
             // Calculate the length of the integer part
-            uLen = (lpaplChar - lpwszOut) - 1;
+            uLen = (UINT) (lpaplChar - lpwszOut) - 1;
 
             // Max the current integer width with this
             lpFmtColStr[aplDimCol].uInts = max (lpFmtColStr[aplDimCol].uInts, uLen);
@@ -1423,13 +1423,13 @@ LPAPLCHAR CompileArrHetero
              && (lpwsz = strchrW (lpwszOut, L'.')))
             {
                 // Calculate the length of the integer part
-                uLen = (lpwsz - lpwszOut);
+                uLen = (UINT) (lpwsz - lpwszOut);
 
                 // Max the current integer width with this
                 lpFmtColStr[aplDimCol].uInts = max (lpFmtColStr[aplDimCol].uInts, uLen);
 
                 // Calculate the length of the fractional part
-                uLen = (lpaplChar - lpwsz) - 1;
+                uLen = (UINT) (lpaplChar - lpwsz) - 1;
 
                 // Max the current fractional width with this
                 lpFmtColStr[aplDimCol].uFrcs = max (lpFmtColStr[aplDimCol].uFrcs, uLen);
@@ -1439,7 +1439,7 @@ LPAPLCHAR CompileArrHetero
                 // We use the following odd construction because
                 //   if the previous SymEntry was a []TCLF,
                 //   lpaplChar EQ lpwszOut
-                uLen = lpaplChar - lpwszOut;
+                uLen = (UINT) (lpaplChar - lpwszOut);
                 uLen = max (uLen, 1) - 1;
 
                 if (IsImmChr (immTypeCur))
@@ -2359,7 +2359,7 @@ LPAPLCHAR FormatArrSimple
                             uLead = 0;
                         else
                         {
-                            uCol = (lpw - lpaplChar) + lpFmtColStr[aplDimCol].uFrcs;
+                            uCol = (UINT) (lpw - lpaplChar) + lpFmtColStr[aplDimCol].uFrcs;
                             uLead = uCmpWid - max (uActLen, uCol);  // # leading blanks to align decimal points
                         } // End IF/ELSE
 
@@ -2370,7 +2370,7 @@ LPAPLCHAR FormatArrSimple
                         // If this is raw output,
                         // break the line if it would exceed []PW
                         //   and the line is non-empty.
-                        uCol = lpwszOut - lpwszOutStart;
+                        uCol = (UINT) (lpwszOut - lpwszOutStart);
                         if (bRawOutput
                          && DEF_INDENT < uCol
                          && uQuadPW < (uLeadBefore + uCmpWid + uCol))
@@ -3351,7 +3351,7 @@ LPPL_YYSTYPE PrimFnDydDownTackJot_EM_YY
                     } // End SWITCH
 
                     // Get the formatted length
-                    uLen = (lpaplChar - lpaplCharIni) - 1;
+                    uLen = (UINT) (lpaplChar - lpaplCharIni) - 1;
 
                     // Check for automatic width
                     if (Auto)
@@ -3461,7 +3461,7 @@ YYALLOC_EXIT:
     // Fill in the result token
     lpYYRes->tkToken.tkFlags.TknType   = TKT_VARARRAY;
 ////lpYYRes->tkToken.tkFlags.ImmType   = 0;     // Already zero from YYAlloc
-////lpYYRes->tkToken.tkFlags.NoDisplay = 0;     // Already zero from YYAlloc
+////lpYYRes->tkToken.tkFlags.NoDisplay = FALSE; // Already zero from YYAlloc
     lpYYRes->tkToken.tkData.tkGlbData  = MakePtrTypeGlb (hGlbRes);
     lpYYRes->tkToken.tkCharIndex       = lptkFunc->tkCharIndex;
